Print totals in quiet mode, if specified
As totals is an explicitly specified parameter, it should also be considered in quiet mode. If a user does not want totals, the parameter can just be left out. Otherwise it is not possible to get totals but suppress output for each file.
As the default for totals is 0, the current behavior if called without -t should not change.